Moved check_channel_counterparty to chain::counterparty
